requirements.txt
streamlit
numpy
pandas
matplotlib
Pillow
torch
torch-geometric
rdkit
